renamed 'graph' -> 'network'
